+++ mi/exception.py
@@ -22,10 +22,12 @@ __all__ = (
     "NotExistRequiredData"
 )
 
+
 class NotExistRequiredData(Exception):
     """
     必要なデータが存在しない場合に発生する例外
     """
+
 
 class NotFoundError(Exception):
     """